+++ index.d.ts
@@ -20,6 +20,7 @@ export interface Config {
   readonly replace?: string;
   readonly fetch?: {
     readonly rewrite?: (path: string) => string;
+    readonly cache?: (path: string, headers: Headers) => string;
     readonly headers?: Headers;
     readonly timeout?: number;
     readonly wait?: number;